refactor: move hook to journeys ui
