+++ libs/journeys/ui/src/libs/useCommonVideoVariantLanguages/useCommonVideoVariantLanguages.mock.ts
